+++ src/main/java/GameCharacters/Character.java
@@ -1,14 +1,19 @@
 package GameCharacters;
 
+import Jobs.Job;
 import Magic.SpellCollection;
+import Races.*;
+
 
 public abstract class Character {
 
     private String name;
     private boolean isAlive;
     private Race race;
-    //private Job job;
+	private Job job;
     protected int level;
+	private int health;
+
     private SpellCollection spellCollection;
 
 
@@ -22,15 +27,15 @@ public abstract class Character {
 
 
 //TODO ändra maxhealth och maxmana
-    public Character(String name, Race race, boolean isAlive) {
+    public Character(String name, Race race,Job job, boolean isAlive, int health) {
         this.name = name;
         this.race = race;
+		this.job = job;
         isAlive = true;
         setStrength(10);
         setIntelligence(10);
         setMaxMana(200);
         remainingMana = maxMana;
-        setMaxHealth(300);
         remainingHealth = maxHealth;
 
     }
@@ -38,10 +43,29 @@ public abstract class Character {
     public String getName() {
         return name;
     }
+    
+	public void increaseHealth(int hp) {
+		if (getHealth() + hp > race.getMaxHealth()) {
+			setHealth(race.getMaxHealth());
+			return;
+		}
+		
+		throw new IllegalStateException();
+	}//increase
 
     public int getLevel(){
         return level;
     }
+    
+    //Lagt health i character
+    
+    public int getHealth(){
+        return health;
+    }
+    
+	protected void setHealth(int health) {
+		this.health = health;
+	}//
 
     public boolean isAlive() {
         return isAlive;
@@ -73,14 +97,6 @@ public abstract class Character {
 
     private void setMaxMana(){
         //Calculate job/race/etc.
-    }
-
-    public int getMaxHealth() {
-        return maxHealth;
-    }
-
-    protected void setMaxHealth(int maxHealth) {
-        this.maxHealth = maxHealth;
     }
 
     public int getRemainingHealth() {
