--- src/main/java/GameCharacters/Human.java
@@ -1,13 +1,0 @@
-package GameCharacters;
-
-
-public class Human extends Race {
-
-    public Human() {
-        setIfCanFly(false);
-        setIfCanSwim(true);
-        setIfCanWalkThroughTerrain(true);
-
-    }
-
-}
