+++ src/main/java/GameCharacters/Player.java
@@ -1,14 +1,16 @@
 package GameCharacters;
 
 import GameCharacters.Character;
+import Jobs.Job;
+import Races.*;
 
 public class Player extends Character {
 
     //private int level;
 
-    public Player(String name, Race race, boolean isAlive) {
+    public Player(String name, Race race,Job job, boolean isAlive, int health) {
 
-        super(name, race, isAlive);
+        super( name,  race, job,  isAlive,  health);
         setLevel(1);
 
     }
