+++ src/main/java/Races/Ogre.java
@@ -1,5 +1,8 @@
 package Races;
 
+import GameCharacters.*;
+
+
 public class Ogre extends Race {
 
 	public Ogre() {
